chore: routes redirection
