+++ index.html
@@ -3,8 +3,8 @@
   <head>
     <meta charset="UTF-8" />
     <link rel="icon" type="image/svg+xml" href="/vite.svg" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Tauri + React + TS</title>
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <title>Kusillo</title>
   </head>
 
   <body>
