+++ package.json
@@ -10,17 +10,22 @@
     "tauri": "tauri"
   },
   "dependencies": {
+    "@emotion/react": "^11.11.1",
+    "@emotion/styled": "^11.11.0",
+    "@mui/material": "^5.13.7",
+    "@tauri-apps/api": "^1.4.0",
+    "formik": "^2.4.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "@tauri-apps/api": "^1.4.0"
+    "react-router-dom": "^6.14.1"
   },
   "devDependencies": {
+    "@tauri-apps/cli": "^1.4.0",
     "@types/node": "^18.7.10",
     "@types/react": "^18.0.15",
     "@types/react-dom": "^18.0.6",
     "@vitejs/plugin-react": "^3.0.0",
     "typescript": "^4.9.5",
-    "vite": "^4.2.1",
-    "@tauri-apps/cli": "^1.4.0"
+    "vite": "^4.2.1"
   }
 }
