+++ src/forms/team_registration.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+export default function SignupForm () {
+
+    return (
+        <h1>
+            Hello sign up form
+        </h1>
+    );
+}
