+++ src/home.tsx
@@ -1,0 +1,41 @@
+import { useState } from "react";
+import reactLogo from "./assets/react.svg";
+import { invoke } from "@tauri-apps/api/tauri";
+import "./App.css";
+import Button from '@mui/material/Button';
+
+function Home() {
+    const [greetMsg, setGreetMsg] = useState("");
+    const [name, setName] = useState("");
+
+    async function greet() {
+        // Learn more about Tauri commands at https://tauri.app/v1/guides/features/command
+        setGreetMsg(await invoke("greet", { name }));
+    }
+
+    return (
+        <div className="container">
+            <h1>Welcome to Kusillo</h1>
+
+            <div className="row">
+                <a href="https://vitejs.dev" target="_blank">
+                    <img src="/vite.svg" className="logo vite" alt="Vite logo" />
+                </a>
+                <a href="https://tauri.app" target="_blank">
+                    <img src="/tauri.svg" className="logo tauri" alt="Tauri logo" />
+                </a>
+                <a href="https://reactjs.org" target="_blank">
+                    <img src={reactLogo} className="logo react" alt="React logo" />
+                </a>
+            </div>
+
+            <p>Rock es cultura</p>
+
+            <div>
+                <Button variant="outlined">You are so smart I love you forever</Button>
+            </div>
+        </div>
+    );
+}
+
+export default Home;
